bot/pyproject.toml
[project]
name = "lms-bot"
version = "0.1.0"
description = "Telegram Bot for LMS"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aiogram>=3.4.0",
    "python-dotenv>=1.0.1",
    "httpx>=0.27.0",
    "aiohttp-socks>=0.11.0",
]
